Mouse: Rework the USB descriptors
Instead of bundling X, Y, and Wheel together, pull out the latter, split it into
resolution modifier + wheel, and base the horizontal wheel on top of that.
Functionally, this should be the same, just structured in a different way.

This seems to upset at least Linux a lot less, and will hopefully fix #18.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>