SAMD: Fix the host wakeup code
We can't include `SAMD21_USBDevice.h`, because it unconditionally redefines
symbols, leading to linking errors. We work this around by implementing pretty
much the same thing its `.wakeupHost()` method does.

This is a workaround, until the SAMD core is fixed to expose the `wakeupHost()`
method in a sane way.

Mind you, this doesn't quite work yet, but at least it compiles, and is a step
in the right direction.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>