HID_::SendReport_: Use a single send on SAMD
On SAMD, we need to send the whole report in one go, otherwise the report does
not arrive to the host. To do this, we merge the report id and the report itself
into a single buffer, and send that instead of sending the report id and the
data separately.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>