Make Keyboard HID report data variables private
These were originally made public at my request in order to (imperfectly) work
around a mid-cycle report problem in Qukeys.  That problem was that, if a plugin
wanted to send one or more extra reports in response to a keyswitch toggling on
or off, it was not possible to safely send such a report from an
`onKeyswitchEvent()` handler, because the new report was likely incomplete.  The
workaround was to give plugins direct access to the old report's data structure,
which could then be modified and sent. This wasn't a good solution, however,
because of potentially complex plugin interactions.

That workaround has since been replaced with a system that only sends reports in
another handler function (`beforeReportingState()` or `afterEachCycle()`).  This
works, and while it creates more complexity, it solves the problem without
requiring direct access to the report data, and better avoids potential bugs.

Signed-off-by: Michael Richters <gedankenexperimenter@gmail.com>