SystemControl: Restore release() and releaseAll()
Since ab9702352f09f4f4ceee0ba639f872d00df06e93,  `SystemControl_::begin()` is a
no-op, calling it in `release()` and `releaseAll()` is not doing what it did
prior to that commit: call `end()`, which is what the release functions really
wanted to do in the first place.

However, calling `end()` is also backwards: `end()` should be calling one of the
release functions instead. So lets do that: `releaseAll()` will now send an
empty report, and `release()` and `end()` will call it.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>