fix NKRO report fencepost
There was a fencepost error in counting the number of bits in the key
bitmap. This resulted in the report descriptor declaring one more Usage
than items in the report, which might cause validation problems on some
HID implementations.

The only affected Usage was Keypad Hexadecimal, which is probably not
used often.

Signed-off-by: Taylor Yu <code@argon.blue>