reset protocol on bus reset
Only reset boot/report protocol on USB bus reset. Windows 10 will
sometimes try to cache report descriptors if it has high confidence
that it's a device it's seen before. This means it won't retrieve
either the configuration descriptor set or the report descriptors when
resetting after receving control from the BIOS/UEFI. The BIOS or UEFI
probably set the boot keyboard to boot protocol, which meant that
when Windows regained control, the boot keyboard was still in boot
protocol, contrary to the USB HID specification.

This requires the higher-level driver to call a polling function,
because detection of USB reset events isn't as portable as checking
the USB device configuration.

Signed-off-by: Taylor Yu <code@argon.blue>